Avoid "overlapping text edits" in "Make Static" refactoring #1045 (#1070)
* Avoid "overlapping text edits" in "Make Static" refactoring #1045

The "Make Static" refactoring currently produces overlapping text edits
in case there are changes in the source code file containing the
refactored method before and after that method. The reason is that the
calls to the refactored method are wrapped into a single MultiTextEdit
whose scope is then overlapping with the TextEdit for the refactored
method itself.

With this fix, the changes to callers of the refactored method are
treated as individual TextEdits rather than one MultiTextEdit, so that
none of the generated text edits have pairwise overlaps. An according
regression test for the behavior is added.

Fixes #1045