Correct "Make Static" refactoring behavior for anonymous classes #1083
The "Make Static" refactoring erroneously interprets a method call
within an anonymous class inside the refactored method as a call on
`this` and thus changes the method invocation to be performed on the
added input parameter of the method. This is (1) semantically incorrect
and (2) leads to compile errors if the type of the added parameter does
not provide the method.

With this fix, the identification of whether a method invocation is
performed on the `this` object of the method to be refactored and not
some instance of an anonymous inner class inside that method is
corrected. For three different scenarios (anonymous class extending
outer class, anonymous class extending other class, and nested anonymous
classes), regression tests are added.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/1083