Perform element changes always in a non UI thread
Currently PackageExplorerContentProvider receives events for changed
elements from different threads and already has handling to perform only
certain actions in the UI thread. But it does not really account for the
case when the event itself is triggered from the UI thread in which case
lengthy operation can take place in the UI.

This now checks if the change is using the UI and otherwise schedules a
job to perform the required actions in the background.