Asynchronously decorate quick outline view via DecorationManager #1922
The Quick Outline View currently applies (i.e., calculates and draws)
decorations for override indicators synchronously. In case the
calculation of the indicator takes long, it blocks the UI.
The ordinary Outline View defers the responsibility of calculating and
applying decorations to the DecorationManager. This ensures that (1)
that task is performed asynchronously via a dedicated job and (2) that
the decoration manager takes care of the current user configuration
(i.e., whether those indicators shall be shown or not), which the Quick
Outline View currently replicates.

With this change, the Quick Outline View uses the same label decorator
as the ordinary Outline View, which makes the DecorationManager
process the override label decorator asynchronously instead of applying
it synchronously.

Contributes to https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/1922