Automate JLS selection toggles in AST view
Rather then explicitly listing versions cycle through the AST supported
versions and create for toggle for each of them.
Mnemonics for these toggles are removed now but they have been broken
already as there have been conflicts (8 and 18, 9 and 19, 10 and 20, 11
and 21, 12 and 22, 13 and 23) rendering mnemonics useless.
This has the good side effect of removing the need to touch AST view for
new Java version support.