Force reconciling if a java project was touched
If a classpath container changes, JDT "touches" the project resulting in
a resource change being triggered. This is currently ignored by the
JavaReconciler leading to possible wrong state displayed in an editor.

This now enhance the JavaReconciler to check for changed java projects
if not already a reconciling was forced.

Fix https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2074