Spare control characters from content assist trigger
On some versions of MacOS, the Ctrl+S hotkey is realized as a key event
with 'S' as character and 'Ctrl' as state mask, rather than the 'XOFF'
character (0x13, also with the 'Ctrl' state mask).

If the auto-activation trigger for Java include the character 's', this
then causes the completion box to show up when the user tries to save
the editor while it's already in a clean state.

This was previously fixed on the platform [1], but needs to be reverted
as generic source viewers might also check for the 'Ctrl' key [2]. In
order to avoid a regression, the check therefore needs to be moved to
the JDT component.

[1] https://github.com/eclipse-platform/eclipse.platform.ui/pull/1556
[2] https://github.com/eclipse-platform/eclipse.platform.ui/issues/2723