Skip import quickfix hierarchy check if base type is unresolved
The commit 546a6978432fa5f2b01083919364848dd89805dc extended the import
check to compare the interfaces of the suggested types with the
interface of the missing type.

In case the base type is unresolved, this then skips all suggestions,
even though it should let all pass so that the user can decide which one
to pick.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2119