Fixed too strict test assumptions in SearchLeakTest
The SearchLeakTest starts search, opens an editor and expects it is
**exactly** `TextEditor.class`.

Now after changes in
https://github.com/eclipse-platform/eclipse.platform.ui/pull/2881 it
finds `ExtensionBasedTextEditor` class, which is a subclass of
`TextEditor`, but test logic doesn't like that.

Fixed test expectation and allow opened editor be a subclass of
`TextEditor`.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2132