Replace pre-generated disabled icon with on-the-fly generated ones
SWT's algorithm to calculate disabled icons on the fly has been enhanced
to produce results that are equal to the existing, pre-generated
disabled icons being used. This makes the usage of pre-generated
disabled icons obsolete. This is particularly important for the
introduction of SVG icons, as the quality of on-the-fly generated
disabled icons based on the original SVGs is better for arbitrary zooms
than the scaled up or scaled downs versions of the PNG-based disabled
icons.

This change removes all explicit specifications of disabled image
descriptors or replaces them with image descriptors derived from their
enabled version.