Do not return null editor input from the ClassFileEditorInputFactory
Currently ClassFileEditorInputFactory return null in cases when
restoring an editor input from an IMemento and the class is not found on
the (current) classpath of the project. This prevents the editor from
showing up and results in a generic error message instead and prevents
proper handling of the case in the editor.

This now in case of an element not found by the project an
InternalClassFileEditorInput is returned as we actually have the class
data there. The editor can then further handel the case much better.