Do not return null editor input from the ClassFileEditorInputFactory
Currently ClassFileEditorInputFactory return null in some cases when
restoring an editor input from an IMemento. While this is legal in
general for IElementFactory it is not very useful for factories that
restore editor inputs because this will fail to instantiate the editor
with a nasty exception without any way for the editor to handle or the
user to understand the underlying problem.

This now changes two things:

1) In case of an error, a special ExceptionalEditorInput is created,
that when later used to gather its underlying data will rethrow the
original exception so we get a useful message about the cause.
2) In case of an element not found by the project an
InternalClassFileEditorInput is returned as we actually have the class
data already there.