If there is no source show the raw class file content
Currently if ClassFileEditor is opened for a classfile that is not on
the classpath it fails with an error "The class file is not on the
classpath" even though it could just display the raw class.

This now adjust the ClassFileEditor in a way to be more lenient and
allows to display class files that are not on the classfile at least
with the default raw view.