Do not return null from ClassFileEditorInputFactory when type not found
Currently the ClassFileEditorInputFactory tries to update a stored
element reference to a type in the project when it encounters an
IOrdinaryClassFile. If that lookup fails (for whatever reason) it return
null that results in a complete crash of the editor during restore as
null is not a valid EditorInput.

This now returns an InternalClassFileEditorInput instead to make it show
the editor again after a restart.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2371