Move the relocation code from the factory to the ClassFileEditor
Currently the ClassFileEditorInputFactory performs some action to
relocate ordinary class files into proper types of the project.

This has some problems:

- recovery should ideally be very fast because this code is called very
early even before the UI is visible
- as it is not supported to return null error handling is not reliable
possible any error will be swallowed and results in a generic error
message that is hard to understand by the user and hard to track down
because there is no stacktrace to the original code

This now do the following:

- create a new carrier input HandleEditorInput that is just a wrapper
around a string and except for the case the memento itself is invalid we
can always return a non-null value
- the existing ClassFileEditor#transformEditorInput method is enhanced
to contain the code previously at the factory

This has the advantage that now we can throw CoreEceptions in case
anything goes wrong. Also the code is now part of the editor itself so
any advanced handling (e.g. delayed loading in a Job, other cases of
transformation needed) can now be handled not only in the editor but
with full access to the editor instance itself.