Restore original type from class file input in background
If ClassFileEditor is restored, underlined class file might be moved in
the meantime (e.g. library on classpath changed version so the type is
still there but the absolute class file path changed).

To check if the type is still on the classpath of the project might
freeze UI for some time, so move this check to a job.

If the job is running very early on startup, project classpath might be
not yet updated (see RequiredPluginsInitializer and UpdateClasspathsJob
in PDE which initialize classpath container lazily). In the case the
type is not found, listen to JDT model changes and refresh the input in
case the project classpath is updated.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2245