Show progress when calculating AST for ClipboardOperationAction #2028 (#2382)
If the copy operation requires an AST, then the AST will be calculated
by forking to a non-UI thread and using the progress service of the
workbench.
This gives the user a chance to cancel the operation while also showing
the current progress.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2028