Make '*' injection in "Open Type" dialog more restrictive #2505 (#2506)
This adapts the modification of the pattern text, that was introduced
with 02958ce874164f2a61fcfba52e975988f3704e89 to only add '*' between
two camel-case words. Otherwise the search might return wrong or
unexpected results when e.g. searching for fully-qualified types.

To test:

1) Search for "java.lang.String". The Java class should be the first
match, as opposed to e.g. "java.lang.AbstractStringBuilder", which is
suggested when searching for "java.lang.*String".

2) Search for "ComAct", which should return the same results as
"Com*Act". More specifically, it should also include types such as the
"ComboBoxActionPropertyChangeListener", which was previously excluded
because the "Box" was not part of the search query.

Closes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2505