Fix types dialog when pattern contains only uppercase letters #2538
This solves a regression where a search pattern, that only contains
upper-case letters, wouldn't show any matches.

The bug was caused with 210f65acc02641c0e26a081aa88ca119c4c2b2ca, where
a fix a different issue was fixed when searching for a type by its fully
qualified name.

Cause of this issue is the inversion of the "canAddAnyStringNext" logic,
which now requires this variable to be initialized with "false".
Otherwise every search string would being with a "*", like "*OOM" for
"OOM".

A test case has been added to cover the relevant cases and to make sure
future changes don't introduce any regressions.

Closes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2538