Propose adding JUnit 6 to classpath when creating a JUnit Jupiter test case
If a Java project has no JUnit container on the classpath,
creating a test case with the respective wizard
will prompt the user to add a JUnit container.

This change adds JUnit 6 to the list of containers in the proposal,
for JUnit Jupiter tests. The list now has JUnit 5 and JUnit 6.

Fixes: #2606