Fix inconsistent project settings regarding java version
The org.eclipse.jdt.junit.runtime.* currently declares a BREE with
Java-11 and therefore also gets this as a JRE11 on the classpath, but in
the build.properties it declares jre.compilation.profile = JavaSE-1.8
and should actually work to run tests on older JREs as well.
Furthermore it has previously declared source/compliance/target level of
1.8 already so it actually was targeted at 1.8 and use that in the
pom.xml as well.

To prevent confusion in an OSGi environment (where it will only resolve
for Java 11+) this now fixes the manifest and classpath to align with
that as well.